replaced half to float.

+++ Assets/Toon(HDRP)/Shaders/UtsLightLoop.hlsl
@@ -484,7 +484,7 @@ float3 UTS_MainLightShadingGrademap(LightLoopContext lightLoopContext, FragInput
 #endif
 
 
-    half shadowAttenuation = lightLoopContext.shadowValue;
+    float shadowAttenuation = lightLoopContext.shadowValue;
 
     DirectLighting lighting = EvaluateBSDF_Directional(lightLoopContext, V, posInput, preLightData, _DirectionalLightDatas[mainLightIndex], bsdfData, builtinData);
 
@@ -496,13 +496,13 @@ float3 UTS_MainLightShadingGrademap(LightLoopContext lightLoopContext, FragInput
 
     float3 defaultLightDirection = normalize(UNITY_MATRIX_V[2].xyz + UNITY_MATRIX_V[1].xyz);
     //v.2.0.5
-    float3 defaultLightColor = saturate(max(half3(0.05, 0.05, 0.05)*_Unlit_Intensity, max(ShadeSH9(half4(0.0, 0.0, 0.0, 1.0)), ShadeSH9(half4(0.0, -1.0, 0.0, 1.0)).rgb)*_Unlit_Intensity));
+    float3 defaultLightColor = saturate(max(float3(0.05, 0.05, 0.05)*_Unlit_Intensity, max(ShadeSH9(float4(0.0, 0.0, 0.0, 1.0)), ShadeSH9(float4(0.0, -1.0, 0.0, 1.0)).rgb)*_Unlit_Intensity));
     float3 customLightDirection = normalize(mul(UNITY_MATRIX_M, float4(((float3(1.0, 0.0, 0.0)*_Offset_X_Axis_BLD * 10) + (float3(0.0, 1.0, 0.0)*_Offset_Y_Axis_BLD * 10) + (float3(0.0, 0.0, -1.0)*lerp(-1.0, 1.0, _Inverse_Z_Axis_BLD))), 0)).xyz);
     float3 lightDirection = normalize(lerp(defaultLightDirection, mainLihgtDirection.xyz, any(mainLihgtDirection.xyz)));
     lightDirection = lerp(lightDirection, customLightDirection, _Is_BLD);
     //v.2.0.5: 
 
-    half3 originalLightColor = mainLightColor.rgb;
+    float3 originalLightColor = mainLightColor.rgb;
 
     float3 lightColor = lerp(max(defaultLightColor, originalLightColor), max(defaultLightColor, saturate(originalLightColor)), _Is_Filter_LightColor);
 
@@ -768,7 +768,7 @@ float3 UTS_MainLight(LightLoopContext lightLoopContext, FragInputs input, int ma
 //DoubleShadeWithFeather
   #endif
 
-    half shadowAttenuation =  lightLoopContext.shadowValue;
+    float shadowAttenuation =  lightLoopContext.shadowValue;
 
     DirectLighting lighting = EvaluateBSDF_Directional(lightLoopContext, V, posInput, preLightData, _DirectionalLightDatas[mainLightIndex], bsdfData, builtinData);
 
@@ -777,11 +777,11 @@ float3 UTS_MainLight(LightLoopContext lightLoopContext, FragInputs input, int ma
 //    float4 tmpColor = EvaluateLight_Directional(context, posInput, _DirectionalLightDatas[mainLightIndex]);
 //    float3 mainLightColor = tmpColor.xyz;
     float3 defaultLightDirection = normalize(UNITY_MATRIX_V[2].xyz + UNITY_MATRIX_V[1].xyz);
-    float3 defaultLightColor = saturate(max(half3(0.05, 0.05, 0.05)*_Unlit_Intensity, max(ShadeSH9(half4(0.0, 0.0, 0.0, 1.0)), ShadeSH9(half4(0.0, -1.0, 0.0, 1.0)).rgb)*_Unlit_Intensity));
+    float3 defaultLightColor = saturate(max(float3(0.05, 0.05, 0.05)*_Unlit_Intensity, max(ShadeSH9(float4(0.0, 0.0, 0.0, 1.0)), ShadeSH9(float4(0.0, -1.0, 0.0, 1.0)).rgb)*_Unlit_Intensity));
     float3 customLightDirection = normalize(mul(UNITY_MATRIX_M, float4(((float3(1.0, 0.0, 0.0)*_Offset_X_Axis_BLD * 10) + (float3(0.0, 1.0, 0.0)*_Offset_Y_Axis_BLD * 10) + (float3(0.0, 0.0, -1.0)*lerp(-1.0, 1.0, _Inverse_Z_Axis_BLD))), 0)).xyz);
     float3 lightDirection = normalize(lerp(defaultLightDirection, mainLihgtDirection.xyz, any(mainLihgtDirection.xyz)));
     lightDirection = lerp(lightDirection, customLightDirection, _Is_BLD);
-    half3 originalLightColor = mainLightColor;
+    float3 originalLightColor = mainLightColor;
 
     float3 lightColor = lerp(max(defaultLightColor, originalLightColor), max(defaultLightColor, saturate(originalLightColor)), _Is_Filter_LightColor);
